split the functions, variables and components within search.js into individual components to improve readability
