+++ src/components/global/variables.js
@@ -1,0 +1,5 @@
+import React from 'react'
+
+const color = '#ffffff'
+
+export { color }
